+++ CHANGELOG.md
@@ -8,7 +8,8 @@
           simply reduce the number of used tags imho
     - [ ] when grouped sort them in the group by name
     - [ ] link each tag in the posts headline to the tag page
-- [ ] group by month
+- [x] group by month
+- [ ] show month-pages and link them
 - [ ] render the schema attributes
 - [ ] add all posts from
   - [x] picostitch